Add support for intrinsic efficiency estimation.
Estimate the angular dependence of the intrinsic efficiency using a ray
tracing approximation and the detection cross sections. Note that this
appears to make the localization accuracy *worse*, we suspect this is
because of the energy dependence of the cross sections, which this
estimate does not account for.

Hence, you probably don't want to actually use this type of detector. I
am committing it for posterity in case someone wants to examine this
effect in the future.